gradio-chatgpt/requirements.txt
openai
langchain
chardet
cchardet
Flask
chromadb
tiktoken
python-dotenv